% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{httpClient}
\alias{httpClient}
\title{Example Asio application}
\usage{
httpClient(host, path)
}
\arguments{
\item{host}{A character variable with an internet host name}

\item{path}{A character variable with an accessible path}
}
\value{
Nothing is returned as the function is invoked for its a side-effect of display
what is retrieved.
}
\description{
This function implements a very simple http client. It follows the \code{http_client.cpp}
example in the Asio documentation but has been adapted to be used as an example function
here.
}
\examples{
\dontrun{
## now commented-out / inactive as the redirect upsets R's own checking
httpClient("www.apache.org", "/robots.txt")
}
}
